no longer nilling tokens on refresh fail
